+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Bugfix where aliases composed of multiple columns would not correctly resolve
 - Allow customization of SQL dialect placeholder format
 - Bugfix to allow forwards and backwards required conditions
+- Add required whitespace to SQL grammars
+- SQL pretty-printer should return everything on one line.
 
 ## 0.2.1 - 7/10/23